fix inApp notification's badge on bell icon
